31-10-2021

+++ basic.cpp
@@ -4,13 +4,12 @@ typedef long long int lli;
 
 int main()
 {
-    int test;
-    cin>>test;
-    for(int e=0;e<test;e++)
-    {
+   lli test;
+   cin>>test;
+   for(lli e=0;e<test;e++)
+   {
 
-    }
+   }
 
     return 0;
 }
-
